cleanup and add update logic

+++ assets/js/main.js
@@ -5,16 +5,17 @@ $(function(){
   // and more complex app, this type of code can get very messy, very fast.
 
   $('#index .delete').on('click', function(){
-    console.log($(this).data('id'));
-
-    $.ajax({
-      url: $(this).data('id'),
-      method: 'DELETE'
-    }).fail(function(err){
-      console.error(err);
-    }).done(function(){
-      location.reload();
-    });
+    var result = confirm('Are you sure?');
+    if (result) {
+      $.ajax({
+        url: $(this).data('id'),
+        method: 'DELETE'
+      }).fail(function(err){
+        console.error(err);
+      }).done(function(){
+        location.reload();
+      });
+    }
 
   });
